update sdk, delete model User.ts and add relation on userCredential to userPhoto
