--- src/shared/models/User.ts
@@ -1,88 +1,0 @@
-/* tslint:disable */
-
-declare var Object: any;
-export interface UserInterface {
-  "realm"?: string;
-  "username"?: string;
-  "email": string;
-  "emailVerified"?: boolean;
-  "id"?: number;
-  "password"?: string;
-  accessTokens?: any[];
-}
-
-export class User implements UserInterface {
-  "realm": string;
-  "username": string;
-  "email": string;
-  "emailVerified": boolean;
-  "id": number;
-  "password": string;
-  accessTokens: any[];
-  constructor(data?: UserInterface) {
-    Object.assign(this, data);
-  }
-  /**
-   * The name of the model represented by this $resource,
-   * i.e. `User`.
-   */
-  public static getModelName() {
-    return "User";
-  }
-  /**
-  * @method factory
-  * @author Jonathan Casarrubias
-  * @license MIT
-  * This method creates an instance of User for dynamic purposes.
-  **/
-  public static factory(data: UserInterface): User{
-    return new User(data);
-  }
-  /**
-  * @method getModelDefinition
-  * @author Julien Ledun
-  * @license MIT
-  * This method returns an object that represents some of the model
-  * definitions.
-  **/
-  public static getModelDefinition() {
-    return {
-      name: 'User',
-      plural: 'Users',
-      path: 'Users',
-      properties: {
-        "realm": {
-          name: 'realm',
-          type: 'string'
-        },
-        "username": {
-          name: 'username',
-          type: 'string'
-        },
-        "email": {
-          name: 'email',
-          type: 'string'
-        },
-        "emailVerified": {
-          name: 'emailVerified',
-          type: 'boolean'
-        },
-        "id": {
-          name: 'id',
-          type: 'number'
-        },
-        "password": {
-          name: 'password',
-          type: 'string'
-        },
-      },
-      relations: {
-        accessTokens: {
-          name: 'accessTokens',
-          type: 'any[]',
-          model: ''
-        },
-      }
-    }
-  }
-}
