+++ src/shared/services/custom/SDKModels.ts
@@ -1,6 +1,5 @@
 /* tslint:disable */
 import { Injectable } from '@angular/core';
-import { User } from '../../models/User';
 import { UserPhoto } from '../../models/UserPhoto';
 import { UserCredential } from '../../models/UserCredential';
 import { StorageSimpleUpload } from '../../models/StorageSimpleUpload';
@@ -11,7 +10,6 @@ export interface Models { [name: string]: any }
 export class SDKModels {
 
   private models: Models = {
-    User: User,
     UserPhoto: UserPhoto,
     UserCredential: UserCredential,
     StorageSimpleUpload: StorageSimpleUpload,
