+++ src/shared/services/custom/index.ts
@@ -1,5 +1,4 @@
 /* tslint:disable */
-export * from './User';
 export * from './UserPhoto';
 export * from './UserCredential';
 export * from './SDKModels';
